+++ app/Http/Controllers/Api/ShareController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\ShareResource;
+use App\Post;
 use App\Share;
 use Illuminate\Http\Request;
 
@@ -16,6 +17,7 @@ class ShareController extends Controller
 
     public function get(Request $request, int $post_id)
     {
+        Share::clearExpired();
         $shares = Share::postIs($post_id)->get();
 
         return response()->json([
@@ -23,8 +25,31 @@ class ShareController extends Controller
         ]);
     }
 
-    public function store(Request $request)
+    public function store(Request $request, int $post_id)
     {
+        $post = Post::findOrFail($post_id);
 
+        $validated = $this->validate($request, [
+            'expiration' => [
+                'required',
+                'in:hour,hours,day,days,week,weeks,month',
+            ]
+        ]);
+
+        $share = new Share(['post_id' => $post->id]);
+        $share->setExpiration($validated['expiration']);
+        $share->generateToken();
+
+        if ($share->save()) {
+            return response()->json([
+                'share' => new ShareResource($share),
+                'status' => 'created',
+            ]);
+        }
+
+        return response()->json([
+            'message' => __('Unable to create link for this content'),
+            'status' => 'error',
+        ]);
     }
 }