+++ app/Share.php
@@ -5,6 +5,7 @@ namespace App;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Support\Carbon;
 use Illuminate\Support\Str;
 
 /**
@@ -13,8 +14,7 @@ use Illuminate\Support\Str;
 class Share extends Model
 {
     protected $fillable = [
-        'sharable_type',
-        'sharable_id',
+        'post_id',
         'token',
         'expires_at',
     ];
@@ -29,7 +29,7 @@ class Share extends Model
 
     public function getUrlAttribute(): string
     {
-        return route('share', $this->token);
+        return route('share', [$this->id, $this->token]);
     }
 
     public function scopePostIs(Builder $query, int $id): Builder
@@ -42,8 +42,48 @@ class Share extends Model
         return $query->where('token', $token);
     }
 
-    public function generateToken()
+    public function setExpiration(string $expiration): self
     {
-        return Str::random(64);
+        $expires = Carbon::now();
+
+        switch ($expiration) {
+            case 'month':
+                $expires->addMonth();
+                break;
+            case 'weeks':
+                $expires->addWeeks(2);
+                break;
+            case 'week':
+                $expires->addWeek();
+                break;
+            case 'days':
+                $expires->addDays(3);
+                break;
+            case 'day':
+                $expires->addDay();
+                break;
+            case 'hours':
+                $expires->addHours(12);
+                break;
+            default:
+            case 'hour':
+                $expires->addHour();
+                break;
+        }
+
+        $this->expires_at = $expires;
+        return $this;
+    }
+
+    public function generateToken(): self
+    {
+        $this->token = Str::random(64);
+        return $this;
+    }
+
+    public static function clearExpired(): void
+    {
+        Share::where('expires_at', '<', Carbon::now()->toDateTimeString())
+            ->delete();
     }
 }