+++ resources/js/components/Modal.vue
@@ -1,7 +1,6 @@
 <template>
     <div class="modal fade"
          :id="id"
-         tabindex="-1"
          role="dialog"
          v-if="show"
     >